requirements.txt
#  Update your requirements.
#  numpy==1.18.1


pillow
comfy
torchvision
torch